Move model files to folders
