+++ server/controllers/sessions_controller.h
@@ -3,8 +3,8 @@
 
 #include <drogon/HttpController.h>
 #include "model/auth/authentication_manager.h"
-#include "model/game_logic_manager.h"
-#include "model/games_storage.h"
+#include "model/logic/game_logic_manager.h"
+#include "model/games/games_storage.h"
 #include "model/sessions/sessions_storage.h"
 
 namespace cavoke::server::controllers {
