+++ server/model/games/games_storage.h
