+++ server/model/logic/game_logic_manager.cpp
