+++ server/model/logic/game_state_storage.h
